Moving media_settings.json functionality to a separate file and updating xcvrd accordingly
